Backend/requirements.txt
asgiref==3.7.2
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
# Pillow==10.1.0
psycopg2==2.9.9
PyJWT==2.8.0
pytz==2023.3.post1
sqlparse==0.4.4
tzdata==2023.3
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pillow==10.2.0
pytesseract==0.3.10
tesseract==0.1.3
